Added Admin, more style, profiles and login-flask

+++ app.py
@@ -1,20 +1,34 @@
 from flask import Flask, render_template, request, redirect, url_for, session, flash
+from werkzeug.utils import secure_filename
 from movieweb_app.data_models import data_models
-# from data_managers.json_data_manager import JSONDataManager
-from movieweb_app.data_models.data_models import db
+from movieweb_app.data_models.data_models import db, User, Movie, Review, UserView, MovieView, ReviewView
 from movieweb_app.data_managers.sql_data_manager import SQLiteDataManager
 from movieweb_app.blueprints.api_routes import api
+from flask_migrate import Migrate
+from werkzeug.security import check_password_hash
+from flask_login import LoginManager, login_user, logout_user, login_required, current_user
+from flask_admin import Admin
 import os
 import secrets
 
 
+# The Api is not put in sql_data_manager
 app = Flask(__name__)
 app.register_blueprint(api, url_prefix='/api')
 app.secret_key = secrets.token_hex(16)
+login_manager = LoginManager(app)
+login_manager.login_view = 'login'
 database_path = os.path.join(app.root_path, 'data', 'data.sqlite')
 app.config['SQLALCHEMY_DATABASE_URI'] = f'sqlite:///{database_path}'
+admin = Admin(app)
 db.init_app(app)
+migrate = Migrate(app, db)
+admin.add_view(UserView(User, db.session))
+admin.add_view(MovieView(Movie, db.session))
+admin.add_view(ReviewView(Review, db.session))
 data_manager = SQLiteDataManager(app)
+
+
 
 @app.errorhandler(400)
 def handle_value_error(error):
@@ -62,6 +76,48 @@ def home():
     return render_template('home.html')
 
 
+@login_manager.user_loader
+def load_user(user_id):
+    return data_manager.find_user_by_id(user_id)
+
+
+@app.route('/users/<int:user_id>/profile', methods=['GET', 'POST'])
+@login_required
+def user_profile(user_id):
+    if current_user.is_authenticated and current_user.user_id == user_id:
+        user = data_manager.find_user_by_id(user_id)
+
+        if request.method == 'POST':
+            action = request.form.get('action')
+            if action == 'update':
+                new_name = request.form.get('new_name')
+                new_password = request.form.get('new_password')
+                profile_picture = request.files['profile_picture']
+                if profile_picture:
+                    # Save the file to a directory
+                    filename = secure_filename(profile_picture.filename)
+                    file_path = os.path.join(os.path.dirname(os.path.abspath(__file__)), 'static\images', filename)
+                    print(file_path)
+                    profile_picture.save(file_path)
+                    data_manager.update_user_profile(user_id, new_name, new_password, filename)
+                    print(user.profile_picture)
+                    flash('Profile picture updated successfully!', 'success')
+                else:
+                    data_manager.update_user_profile(user_id, new_name, new_password)
+                    flash('User profile updated successfully!', 'success')
+            elif action == 'delete':
+                db.session.delete(user)
+                db.session.commit()
+                logout_user()
+                flash('User deleted successfully!', 'success')
+                return redirect(url_for('home'))
+            else:
+                flash('Invalid action', 'error')
+
+        return render_template('user_profile.html', user=user)
+    return redirect(url_for('home'))
+
+
 @app.route('/login', methods=['POST'])
 def login():
     """
@@ -69,24 +125,33 @@ def login():
     Returns:
         flask.Response: The rendered template based on the login result.
     """
+    if current_user.is_authenticated:
+        # User is already logged in, handle as needed
+        return redirect(url_for('user_movies', user_id=current_user.user_id))
     username = request.form.get('username')
     password = request.form.get('password')
-    user_id = data_manager.login(username, password)
-    if user_id:
-        session['user_id'] = user_id
-        return redirect(url_for('user_movies', user_id=user_id))
+    user = data_manager.find_user_by_name(username)
+    print("Searching for user with name:", username)
+    print("Found user:", user)
+    if user and check_password_hash(user.password, password):
+        login_user(user)
+        next_url = request.args.get('next')
+        if not next_url or not next_url.startswith('/users/'):
+            next_url = url_for('user_movies', user_id=user.user_id)
+        return redirect(next_url)
     else:
         return render_template('home.html', error='Invalid credentials', username=username)
 
 
 @app.route('/logout')
+@login_required
 def logout():
     """
     Logout route.
     Returns:
         flask.Response: A redirect response to the home page.
     """
-    session.pop('user_id', None)
+    logout_user()
     return redirect(url_for('home'))
 
 
@@ -108,29 +173,32 @@ def user_authorization(user_id):
     Returns:
         flask.Response: The rendered template based on the authorization result.
     """
+    user = data_manager.find_user_by_id(user_id)
+
     if request.method == 'POST':
         password = request.form.get('password')
-        user = data_manager.find_user_by_id(user_id)
-        if user and data_manager.check_user_password(user_id, password):
+
+        if user and check_password_hash(user.password, password):
             session['user_id'] = user_id
             return redirect(url_for('user_movies', user_id=user_id))
         else:
-            return render_template('users.html', users=data_manager.get_all_users_names(), error='Invalid password')
+            flash('Invalid password', 'error')
+
+    # If the request method is GET or the password is invalid, or any other error occurs.
     return render_template('users.html', users=data_manager.get_all_users_names())
 
 
-@app.route('/users/<user_id>/movies')
+@app.route('/users/<int:user_id>/movies')
+@login_required
 def user_movies(user_id):
     """
     User movies route.
     Returns:
         flask.Response: The rendered template for the user's movies.
     """
-    if 'user_id' in session and session['user_id'] == user_id:
-        user = data_manager.find_user_by_id(user_id)
-        if user:
-            movies = data_manager.get_user_movies(user_id)
-            return render_template('movies.html', movies=movies, user_id=user_id)
+    if current_user.is_authenticated and current_user.user_id == user_id:
+        movies = data_manager.get_user_movies(user_id)
+        return render_template('movies.html', movies=movies, user_id=user_id)
     return render_template('404.html', error='User not found'), 404
 
 
@@ -150,6 +218,7 @@ def add_user():
 
 
 @app.route('/users/<user_id>/recommend_movie', methods=['GET'])
+@login_required
 def recommend_movie(user_id):
     user = data_manager.find_user_by_id(user_id)
     try:
@@ -167,30 +236,33 @@ def recommend_movie(user_id):
         return f"Error {e}"
 
 
-@app.route('/users/<user_id>/add_movie', methods=['GET', 'POST'])
+@app.route('/users/<int:user_id>/add_movie', methods=['GET', 'POST'])
+@login_required
 def add_movies(user_id):
     """
     Add movie route.
     Returns:
         flask.Response: The rendered template for adding a movie or a redirect response after adding a movie.
     """
-    if 'user_id' in session and session['user_id'] == user_id:
+    if current_user.is_authenticated and current_user.user_id == user_id:
         if request.method == 'POST':
             name = request.form.get('name')
             data_manager.add_user_movie(user_id, name)
-            return redirect(url_for('user_movies', user_id=user_id)), 201
+            flash('Movie added successfully!', 'success')
+            return redirect(url_for('user_movies', user_id=user_id)), 301
         return render_template('add_movie.html', user_id=user_id)
     return redirect(url_for('home'))
 
 
-@app.route('/users/<user_id>/update_movie/<movie_id>', methods=['GET', 'POST'])
+@app.route('/users/<int:user_id>/update_movie/<movie_id>', methods=['GET', 'POST'])
+@login_required
 def update_movie(user_id, movie_id):
     """
     Update movie route.
     Returns:
         flask.Response: The rendered template for updating a movie or a redirect response after updating a movie.
     """
-    if 'user_id' in session and session['user_id'] == user_id:
+    if current_user.is_authenticated and current_user.user_id == user_id:
         if request.method == 'POST':
             try:
                 name = request.form.get('name')
@@ -206,14 +278,15 @@ def update_movie(user_id, movie_id):
     return redirect(url_for('home'))
 
 
-@app.route('/users/<user_id>/delete_movie/<movie_id>', methods=['GET', 'POST'])
+@app.route('/users/<int:user_id>/delete_movie/<movie_id>', methods=['GET', 'POST'])
+@login_required
 def delete_movie(user_id, movie_id):
     """
     Delete movie route.
     Returns:
         flask.Response: The rendered template for deleting a movie or a redirect response after deleting a movie.
     """
-    if 'user_id' in session and session['user_id'] == user_id:
+    if current_user.is_authenticated and current_user.user_id == user_id:
         movie = data_manager.get_movie_by_id(user_id, movie_id)
         if movie:
             if request.method == 'POST':
@@ -223,19 +296,33 @@ def delete_movie(user_id, movie_id):
     return redirect(url_for('home'))
 
 
-@app.route('/users/<user_id>/delete_review/<review_id>', methods=['GET', 'POST'])
+@app.route('/users/<int:user_id>/delete_review/<review_id>', methods=['GET', 'POST'])
+@login_required
 def delete_review(user_id, review_id):
-    if 'user_id' in session and session['user_id'] == user_id:
-        review = data_manager.get_review_by_id(review_id)
-        if review:
-            if request.method == 'POST':
-                data_manager.delete_review(user_id, review_id)
-                return redirect(url_for('user_movies', user_id=user_id))
-            return render_template('delete_review.html', user_id=user_id, review_id=review_id)
+    print(f"Delete Review: user_id={user_id}, review_id={review_id}")
+
+    # Check if the user is authorized to delete this review
+    review = data_manager.get_review_by_id(review_id)
+    print("User is authenticated:", current_user.is_authenticated)
+    if review:
+        print(f"Review user_id={review.user.user_id}")
+    print(type(current_user.user_id))
+    print(type(review.user.user_id))
+    print(type(user_id))
+
+    if current_user.is_authenticated and current_user.user_id == user_id and review and review.user.user_id == user_id:
+
+        if request.method == 'POST':
+            data_manager.delete_review(user_id, review_id)
+            return redirect(url_for('movie_detail', user_id=user_id, movie_id=review.movie_id))
+        return render_template('delete_review.html', user_id=user_id, review_id=review_id)
+
+    print("Authorization failed")
     return redirect(url_for('home'))
 
 
 @app.route('/<user_id>/add_review/<movie_id>', methods=['GET', 'POST'])
+@login_required
 def add_review(user_id, movie_id):
     if request.method == 'POST':
         review_text = request.form.get('review_text')
@@ -262,6 +349,7 @@ def add_review(user_id, movie_id):
 
 
 @app.route('/<user_id>/movie_detail/<movie_id>', methods=['GET', 'POST'])
+@login_required
 def movie_detail(user_id, movie_id):
     """
     Movie detail route.
@@ -275,9 +363,6 @@ def movie_detail(user_id, movie_id):
         all_users = data_manager.get_all_users_names()
         return render_template('movie_detail.html', user=user, movie=movie, reviews=reviews, all_users=all_users)
     return render_template('404.html', error='Movie not found'), 404
-
-
-
 
 
 if __name__ == '__main__':
